requirements.txt
fitz==0.0.1.dev2
langchain==0.3.25
langchain_core==0.3.64
langchain_groq==0.3.2
langgraph==0.4.8
openai_whisper==20240930
Pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.1.0
